==> Jupyter_Notebook_files/Step_01_Dataset_and_Data_Collection.ipynb <==
# Download data

For my captone project, I downloaded the dataset from Yelp (https://www.yelp.com/dataset) in August 2019. The datasets were prepared by Yelp in JSON format to let students use their data for their projects (https://www.yelp.com/dataset/challenge). The dataset is stored in 5 files of JSON format. Each file was composed of a single object type - one-JSON-object per line.

# About files

I loaded and transformed Yelp data challenge datasets into Pandas DataFrames:
- **Users** include user information from registered users on Yelp. It contains reviews per user, number of useful reviews they posted, type of reviews, etc.
- **Review** include reviews submitted by customers about the business
- **Checkin** include check-in times of customers at given business store
- **Tip** include tips per business id
- **Business** contain attributes of business listed on Yelp such as categories (e.g. good for lunch), city info, hours, location and Yelp stars. These columns are consistent with Yelp.com. 

# Import packages

import json
import warnings
warnings.filterwarnings('ignore')
from copy import deepcopy
import pickle
import pandas as pd
import numpy as np
from sklearn.feature_extraction import DictVectorizer

# Turn business.json to a Pandas Dataframe

file = '/Users/mai/Desktop/yelp_dataset/yelp_business.json'

# Decode JSON
with open(file) as f:
    data = [json.loads(line) for line in f]
df = pd.DataFrame(data)

# Check how the dataframe looks like
df.head()

# Drop unnecessary columns for this project
df.drop(columns=['address', 'hours', 'postal_code'], inplace=True)

# Check how the nested dictionary in 'attributes' column looks like
# Check the first item
df['attributes'][0]

# Check the inside of nested dictionary - GoodForMeal
df['attributes'][0]['GoodForMeal']

# Check the inside of nested dictionary - BusinessParking
df['attributes'][0]['BusinessParking']

# Compare the df['attributes'][1] with others
# Each business_id contains different information in the attributes column
print('Example #1', df['attributes'][4])
print('----------')
print('Example #2', df['attributes'][1000])

#  Expand the attributes and category columns by using DictVectorizer()

Since each business contains different amount of attribute information in the 'attribute' column, I decided to expand the columns by using  DictVectorizer(). However, the following columns needs to be manually expanded due to the dictionary type for these keys: Ambience, Business Parking, Good For Meal. I set DictVectorizer() for each keys to avoid memory errors and overwriting information. I also made the data to a list since .fit() and .transform() which only accepts a lit of dictionaries.

# Remove empty cells in 'attributes' column in df
my_dict = df[['attributes']].dropna()

# Remove empty cells in my_dict
my_dict.dropna(inplace=True)

# Check shape
print(my_dict.shape)

# Check how the attribute column looks like
my_dict.head()

# Create extra columns indicating the presence or absence of a category with a value of 1 or 0
dvec_attributes = DictVectorizer()
dmat_attributes = dvec_attributes.fit_transform(my_dict['attributes'])
df_attributes = pd.DataFrame(dmat_attributes.toarray(), columns=dvec_attributes.get_feature_names(), index=my_dict['attributes'].index)
df_attributes.shape

# Drop unnecessary columns
# 'BusinessParking', 'GoodForMeal', 'Ambience': I will apply DictVectorizer() separately thus this dataframe doesn't need to keep the columns here
# DietaryRestrictions / BestNights: I found this does not contain insightful data
# HairSpecializesIn: I focus on restaurants thus this has to be removed
# =u: Original Business.JSON has duplicated keys in the nested dictionary. These contain =u

remove_keys = ['BusinessParking', 'GoodForMeal', 'Ambience', 'Music',\
               'DietaryRestrictions', 'BestNights', 'HairSpecializesIn', '=u']
remove_col_list = []
i = 0
for i in range(len(remove_keys)):
    for col in df_attributes.columns:
        if remove_keys[i] in col:
            remove_col_list.append(col)
            i + 1
print(remove_col_list[1])
print(len(remove_col_list))

# Remove all the columns listed in remove_col_list from the DataFrame
df_attributes.drop(columns=remove_col_list, inplace=True)
df_attributes.shape

# Create columns per ambience feature
# Fill NaN in the empty cells
my_dict['ambience'] = my_dict['attributes'].map(lambda x: (x['Ambience']) if 'Ambience' in x else np.nan)

# Delete keys from 'attributes' column to avoid duplicates
selected_key = 'Ambience'
def dropkey(x, key=selected_key):
    y = deepcopy(x)
    if key in y:
         del y[key]
    return y

# Apply a function to all the items in the list
my_dict['attributes'] = my_dict['attributes'].map(dropkey)

# Replace 'None' with NaN in the column
my_dict['ambience'][my_dict['ambience'] == 'None'] = np.nan

# Apply a function to all the items in the list
my_dict['ambience'] = my_dict['ambience'].map(lambda x: eval(x) if type(x)==str else x)

# Remove all NaN values before using DictVectorizer()
my_dict.dropna(inplace=True)

# Check the shape
my_dict.shape

# Apply DictVectorizer to extract features and put them into a Pandas DataFrame
dvec_ambience = DictVectorizer()
dmat_ambience = dvec_ambience.fit_transform(my_dict['ambience'])

# Create a DataFrame
df_ambience = pd.DataFrame(dmat_ambience.toarray(), columns=dvec_ambience.get_feature_names(), index=my_dict['ambience'].index)

# Check how the ambience features were expanded to columns
df_ambience.columns

# Change column names for future use
df_ambience.columns = [
                        'ambience_Casual',
                        'ambience_Classy',
                        'ambience_Divey',
                        'ambience_Hipster',
                        'ambience_Intimate',
                        'ambience_Romantic',
                        'ambience_Touristy',
                        'ambience_Trendy',
                        'ambience_Upscale',
                      ]

# Check how the DataFrame looks like
df_ambience.head()

# Check the shape
df_ambience.shape

## Create columns per Business Parking features

# Create columns per Business Parking features
# Fill NaN in the empty cells
my_dict['businessparking'] = my_dict['attributes'].map(lambda x: (x['BusinessParking']) if 'BusinessParking' in x else np.nan)
my_dict['businessparking'].head()

# Follow the same step as Ambience features
my_dict['attributes'] = my_dict['attributes'].map(dropkey)
my_dict['businessparking'][my_dict['businessparking'] == 'None'] = np.nan
my_dict['businessparking'] = my_dict['businessparking'].map(lambda x: eval(x) if type(x)==str else x)

# Apply DictVectorizer to extract features and put them into a Pandas dataframe
dvec_parking = DictVectorizer()
dmat_parking = dvec_parking.fit_transform(my_dict['businessparking'].fillna(method='ffill'))
df_parking = pd.DataFrame(dmat_parking.toarray(), columns=dvec_parking.get_feature_names(), index=my_dict['businessparking'].index)

# Change column names for modelling and analysis
df_parking.columns = [
                        'parking_Garage',
                        'parking_Lot',
                        'parking_Street',
                        'parking_Valet',
                        'parking_Validated'
                      ]

df_parking.head()

## Create columns per GoodForMeal features

# Fill NaN in the empty cells
my_dict['goodformeal'] = my_dict['attributes'].map(lambda x: (x['GoodForMeal']) if 'GoodForMeal' in x else np.nan)

# Follow the same step as Ambience features
my_dict['attributes'] = my_dict['attributes'].map(dropkey)
my_dict['goodformeal'][my_dict['goodformeal'] == 'None'] = np.nan
my_dict['goodformeal'] = my_dict['goodformeal'].map(lambda x: eval(x) if type(x)==str else x)
#my_dict['GoodForMeal'].dropna(inplace=True)

# Apply DictVectorizer to extract features and put them into a Pandas dataframe
dvec_meal = DictVectorizer()
dmat_meal = dmat_meal = dvec_meal.fit_transform(my_dict['goodformeal'].fillna(method='ffill'))
df_meal = pd.DataFrame(dmat_meal.toarray(), columns=dvec_meal.get_feature_names(), index=my_dict['goodformeal'].index)

# Change column names for modelling and analysis
df_meal.columns = [
                        'meal_Breakfast',
                        'meal_Brunch',
                        'meal_Dessert',
                        'meal_Dinner',
                        'meal_Latenight',
                        'meal_Lunch'
                      ]

df_meal.head()

# Check the DataFrame shape
df_meal.shape

## Merge all dataframes created from business.json

businessdf = pd.concat([df, df_attributes, df_ambience, df_meal, df_parking], axis=1)

# Drop the original column
businessdf.drop(columns='attributes', inplace=True) 
businessdf.head(2)

# Replace NaN values in the dataframe
businessdf = businessdf.fillna(0)

## Clean up the column names before I save the dataframe

The business.json file did not have clear dictionary keys, thus I decided to rename the column keys for future use.

# Rename columns since these include {}, =, ' in the name (e.g. NoiseLevel='average')
businessdf.columns = businessdf.columns.str.replace("{", "")
businessdf.columns = businessdf.columns.str.replace("}", "")
businessdf.columns = businessdf.columns.str.replace("=", "_")
businessdf.columns = businessdf.columns.str.replace("'", "_")

# Replace none to Unknown
businessdf.columns = businessdf.columns.str.replace('none', 'unknown')
businessdf.columns = businessdf.columns.str.replace('byob', 'bring_your_own_bottle')

# Lowercase column names
businessdf.columns = businessdf.columns.map(lambda x: x.lower())

# Remove 'restaurants' from column names
businessdf.columns = businessdf.columns.map(lambda x: x.replace('restaurants', ''))

 # Rename column name clearly
businessdf.rename(columns={
     'is_open': 'business_status','name': 'business_name',
     'review_count': 'review_count_per_business', 'stars': 'average_stars_per_business', 'state': 'US_state',
     'alcohol__beer_and_wine_': 'alcohol_beer_and_wine', 'alcohol__full_bar_': 'alcohol_full_bar', 'alcohol__unknown_': 'alcohol_not_available',
     'bring_your_own_bottlecorkage__no_': 'bring_your_own_bottle_corkage_no',
     'bring_your_own_bottlecorkage__yes_corkage_': 'bring_your_own_bottle_corkage_yes_corkage',
     'bring_your_own_bottlecorkage__yes_free_': 'bring_your_own_bottle_corkage_yes_free',
     'bring_your_own_bottlecorkage_unknown': 'bring_your_own_bottle_corkage_unknown',
     'hastv_false': 'hasTV_false', 'hastv_unknown': 'hasTV_unknown', 'hastv_true': 'hasTV_true',
     'noiselevel__average_': 'noiselevel_average', 'noiselevel__loud_': 'noiselevel_loud',
     'noiselevel__quiet_': 'noiselevel_quiet', 'noiselevel__very_loud_': 'noiselevel_very_loud',    
     'attire__casual_': 'dresscode_casual', 'attire__dressy_': 'dresscode_dressy',
     'attire__formal_': 'dresscode_formal', 'attire_unknown': 'dresscode_unknown',
     'pricerange2_1': 'spricerange_1',
     'pricerange2_2': 'pricerange_2', 'pricerange2_3': 'pricerange_3',
     'pricerange2_4': 'pricerange_4', 'pricerange2_unknown': 'pricerange_unknown',
     'smoking__no_': 'smoking_no', 'smoking__outdoor_': 'smoking_outdoor', 'smoking__yes_': 'smoking_yes',
     'wifi__free_': 'wifi_free', 'Wifi__no_': 'wifi_no',
     'wifi__paid_': 'wifi_paid'}, inplace=True)

# Check how the DataFrame looks like
businessdf.head()

## Drop one column per attribute group 

 To maximize the performance of models, I decided to drop a column. 
 For example, There were 'BikeParking = True/False/None'. I removed 'BikeParking=False' because when 'BikeParking=True' has 1, 'Bikeparking=True'=0 means 'BikeParking=False'

# List up columns to drop
drop_col_list = []

for col in businessdf.columns:
    if 'false' in col:
        drop_col_list.append(col)
print(drop_col_list)

# Drop the listed columns from the DataFrame
businessdf.drop(columns=drop_col_list, inplace=True)

# Pickle results 
businessdf.to_pickle('businessdf.pickle')

# Check if I successfully pickled the data
pickled_businessdf = pd.read_pickle('/Users/mai/Desktop/yelp_dataset/to_submit/businessdf.pickle')
pickled_businessdf.head()

# Check the review file looks like

reviewjson = '/Users/mai/Desktop/yelp_dataset/review.json'

# Decode JSON
with open(reviewjson) as f:
    reviewdata = [json.loads(line) for line in f]
reviewdf = pd.DataFrame(reviewdata)

# Pickle results 
reviewdf.to_pickle('reviewdf.pickle')

# Check if I successfully pickled the data
pickled_reviewdf = pd.read_pickle('/Users/mai/Desktop/yelp_dataset/to_submit/reviewdf.pickle')
pickled_reviewdf.head()

# Check the user file looks like

userjson = '/Users/mai/Desktop/yelp_dataset/user.json'

# Decode JSON
with open(userjson) as f:
    userdata = [json.loads(line) for line in f]
userdf = pd.DataFrame(userdata)

# Pickle results 
userdf.to_pickle('userdf.pickle')

# Check if I successfully pickled the data
pickled_userdf = pd.read_pickle('/Users/mai/Desktop/yelp_dataset/to_submit/userdf.pickle')
pickled_userdf.head()

# Check the checkin file looks like

checkinjson = '/Users/mai/Desktop/yelp_dataset/checkin.json'

# Decode JSON
with open(checkinjson) as f:
    checkindata = [json.loads(line) for line in f]
checkindf = pd.DataFrame(checkindata)

# Pickle results 
checkindf.to_pickle('checkindf.pickle')

# Check if I successfully pickled the data
pickled_checkindf = pd.read_pickle('/Users/mai/Desktop/yelp_dataset/to_submit/checkindf.pickle')
pickled_checkindf.head()

# Check the tip file looks like

tipjson = '/Users/mai/Desktop/yelp_dataset/tip.json'

# Decode JSON
with open(tipjson) as f:
    tipdata = [json.loads(line) for line in f]
tipdf = pd.DataFrame(tipdata)

# Pickle results 
tipdf.to_pickle('tipdf.pickle')

# Check if I successfully pickled the data
pickled_tipdf = pd.read_pickle('/Users/mai/Desktop/yelp_dataset/to_submit/tipdf.pickle')
pickled_tipdf.head()
