==> Jupyter_Notebook_files/Step_02_Data_Cleaning.ipynb <==
# Import packages

import warnings
warnings.filterwarnings('ignore')
from copy import deepcopy
import pickle
import pandas as pd
import numpy as np

# Load business data

df = pd.read_pickle('/Users/mai/Desktop/yelp_dataset/to_submit/businessdf.pickle')
df.head()

# Extract Restaurant or food store information

# Specify rows data that contains 'Restaurants' or 'Food' in the 
restaurants = df[df['categories'].str.contains('Restaurant|Food') == True]

# Check the shape
restaurants.shape

# Check if I extracted restaurant and food data successfully 
restaurants.categories[:5]

# Put labels to business information for future use
Business_type = ['Restaurants', 'Food']
for b in Business_type:
    restaurants.loc[restaurants['categories'].str.contains(b),'business_type'] = b

# Those with NaN values after the above step are 'Restaurant Supplies', thus, I decided to drop the information
restaurants['business_type'].dropna(inplace=True)

# Create a column of couisine information

# Create a cuisine column based on a popular cuisine list
cuisines = ['African', 'Algerian', 'American', 'Belgian', 'Brazilian', 'British', \
            'Cajun', 'Canadian', 'Chinese', 'Cuban', 'Egyptian', 'French', \
            'German', 'Greek', 'Haitian', 'Hawaiian', 'Indian', 'Irish', 'Italian', 
            'Japanese', 'Jewish', 'Kenyan', 'Korean', 'Latvian', \
            'Libyan', 'Mediterranean', 'Mexican', 'Middle Eastern', 'Mormon', \
            'Nigerian', 'Peruvian', 'Polish', 'Portuguese', 'Russian', \
            'Salvadorian', 'Scottish', 'Spanish', 'Swedish', \
            'Tahitian', 'Thai', 'Tibetan', 'Turkish', 'Vietnamese','Welsh'\
            'Venezuelan', 'Argentine', 'Cantonese', 'Salvadoran', 'Arabian',\
            'Moroccan', 'Russian', 'Uzbek', 'Armenian', 'Lebanese', 'Indonesian', \
            'Singaporean']

# Find a cuisine name in the 'categories' column and create a new colomn (or fill 1 in the relevant column)
# E.g. When a cell contains 'American', it creates an 'American' column or fill 1 in the 'American' column if it exists already
for c in cuisines:
    restaurants.loc[restaurants['categories'].str.contains(c),'cuisine'] = c

# Create a column of food type information

# Create a food type column based on a popular cuisine list
Food_type = ['Arabian', 'Armenian', 'Asian Fusion', 'Barbeque', 'Basque', 
            'Breakfast & Brunch', 'Buffets', 'Burgers', 'Cafes', \
            'Cafeteria', 'Cajun', 'Cheesesteaks', 'Chicken Shops', \
            'Chicken Wings', 'Coffee & Tea', 'Comfort Food', 'Creperies', \
            'Delis', 'Dim Sum', 'Diners', 'Falafel', 'Fast Food', 'Fish & Chips', \
            'Fondue', 'Food Court', 'Food Stands', 'Gastropubs', 'Halal', \
            'Hot Dogs', 'Hot Pot', 'Kebab', 'Kosher', 'Moroccan', 'Noodles',\
            'Pizza', 'Ramen', 'Russian', 'Salad', 'Salvadoran', 'Sandwiches', \
            'Seafood', 'Soup', 'Steakhouses', 'Sushi Bars', 'Tacos', 'Tapas Bars', \
            'Tapas Bars', 'Uzbek', 'Waffles', 'Wraps', ]

# Do the same as 'cuisine' columns
for f in Food_type:
    restaurants.loc[restaurants['categories'].str.contains(f),'food_type'] = f

# Dummify columns 

dummy_business_type = pd.get_dummies(restaurants[['business_type']], drop_first=True)

dummy_cuisine_and_food = pd.get_dummies(restaurants[['cuisine', 'food_type']])

restaurants_ver2 = pd.concat([restaurants, dummy_business_type, dummy_cuisine_and_food], axis=1)

restaurants_ver2 = restaurants_ver2.drop(['business_type', 'categories', 'cuisine', 'food_type'], axis=1)

# Feature generation 

## Create a column for total checkin counts

# Load check-in data
checkins = pd.read_pickle('/Users/mai/Desktop/yelp_dataset/to_submit/checkindf.pickle')
checkins.head()

# Create a new DataFrame to hold the timeseries of the checkins
df_checkins = pd.DataFrame()

for id in range(len(checkins.date)):
    
    df_checkins_id = pd.DataFrame()
    
    split = checkins.date[id].split(',')
    df_checkins_id['Checkin_date'] = pd.to_datetime(split)
    df_checkins_id['Business_id'] = checkins.business_id[id]
    df_checkins_id['Year'] = df_checkins_id['checkin_date'].dt.year
    

    df_checkins_id = df_checkins_id.groupby(['business_id','year']).agg(['count']).reset_index()
#    print(df_checkins_id)

    df_checkins = pd.concat([df_checkins, df_checkins_id])

# Pickle results to avoid running all the above code again - it took hours to finish running
df_checkins.to_pickle('df_checkins.pickle')

# Load data
df_checkins = pd.read_pickle('/Users/mai/Desktop/yelp_dataset/to_submit/df_checkins.pickle')
df_checkins.head()

# Check the DataFrame shape
df_checkins.shape

# Rename the column to make it more clear
df_checkins.rename(columns={'number_checkins': 'checkins_peryear'}, inplace=True)

# Current DataFrame structure: one check-in information per row
# Group the checkin information by business_id
total_checkins = pd.DataFrame(df_checkins.groupby('business_id')['checkins_peryear'].sum()).reset_index()

# Estimate the number of years the business is listed on Yelp
total_checkins_yearcount = pd.DataFrame(df_checkins.groupby('business_id')['year'].count()).reset_index()

# Put all data in a DataFrame
df_checkins_ver2 = total_checkins.merge(total_checkins_yearcount, on='business_id')
df_checkins_ver2.columns=['business_id', 'total_checkins', 'total_checkinyear_count']
df_checkins_ver2.head()

# Put all data in a DataFrame
restaurants_ver3 = df_checkins_ver2.merge(restaurants_ver2, on='business_id')

## Create a column for estimated operating years

# Create a copy to avoid overwrting
df_operatingyears = checkins.copy()

# Check how the DataFrame looks like
df_operatingyears[['business_id', 'date']].head()

# Find the first review year

minyear = []

def minyear_extract(x):
    data = x.split(',')
    data = [int(x.strip()[:4]) for x in data]
    minyear.append(min(data))
    return min(data)

# Create a column that contains first review year information
df_operatingyears['minyear'] = df_operatingyears.date.apply(minyear_extract)

# Find the latest review year

maxyear = []

def maxyear_extract(x):
    data = x.split(',')
    data = [int(x.strip()[:4]) for x in data]
    maxyear.append(max(data))
    return max(data)

# Create a column that contains latest review year information
df_operatingyears['maxyear'] = df_operatingyears.date.apply(maxyear_extract)

# Estimate the operating years 
df_operatingyears['estimated_opearating_years'] = df_operatingyears['maxyear'] - df_operatingyears['minyear']

# Put all data in a DataFrame
restaurants_ver4 = df_operatingyears.merge(p_restaurants_ver3, on='business_id')

# Drop the date column since I don't need to use it again
restaurants_ver4.drop(columns='date', inplace=True)

## Create a column for total review count

# Load review data
reviewdf = pd.read_pickle('/Users/mai/Desktop/yelp_dataset/to_submit/reviewdf.pickle')
reviewdf.head(2)

# Group reviews by business_id
total_reviews = pd.DataFrame(reviewdf.groupby('business_id').count()['review_id'])
total_reviews = total_reviews.reset_index()

# Put all data in a DataFrame
restaurants_ver5 = total_reviews.merge(restaurants_ver4, on='business_id')

# Create predictor variables

# Create categories per business type
# Businesses that is active and run for 3+ years: Successful
# Businesses that is active but run for less than 3 years: Active for less than three years
# Busienss that is closed: Failed
predictions = []

for o in range(len(restaurants_ver5)):
    if (restaurants_ver5['estimated_opearating_years'][o] >= 3) & (restaurants_ver5['business_status'][o] == 1):
        predictions.append('successful')
    elif (restaurants_ver5['estimated_opearating_years'][o] < 3) & (restaurants_ver5['business_status'][o] == 1):
        predictions.append('active for less than three years')
    else:
        predictions.append('failed')

# Create a new column
restaurants_ver5['prediction_col'] = predictions

# Check the distribution
total = len(restaurants_ver5['prediction_col'])
successful_count = len(restaurants_ver5[restaurants_ver5['prediction_col'] == 'successful'])
active_count = len(restaurants_ver5[restaurants_ver5['prediction_col'] == 'active for less than three years'])
failed_count = len(restaurants_ver5[restaurants_ver5['prediction_col'] == 'failed'])

# Remove data that are not necessary for this capstone
# This capstone predicts 'successful' or 'failed', thus other information need to be removed
restaurants_clean = restaurants_ver5[restaurants_ver5['prediction_col'] != 'active for less than three years']

# Pickle results to avoid running all cells 
restaurants_clean.to_pickle('restaurants_clean.pickle')

# Load data
p_restaurants_clean = pd.read_pickle('/Users/mai/Desktop/yelp_dataset/to_submit/restaurants_clean.pickle')
p_restaurants_clean.head()
